Require pjcdawkins's fork of guzzle-oauth2-plugin
The fork contains bug fixes and features to an otherwise legacy project. This
used a patch workflow before (based on cweagans/composer-patches), but there
wasn't a reproducible way to update patches.

This commit also sets the "platform" to PHP 5.5.9, so dependencies won't be
accidentally upgraded to a point where compatibility with old PHP versions is
lost.

The 2.x version of this library will not support end-of-life PHP versions.